fixed floodfill memory limitation

+++ floodfill.h
@@ -2,4 +2,4 @@
 #include "queue.h"
 
 //returns the length of the largest cluster
-extern int findLargestCluster(char** array, int size);
+extern unsigned long long findLargestCluster(char** array, int size);
